+++ golos-ios/Common/CoreData/Entities/User+CoreDataProperties.swift
@@ -33,8 +33,8 @@ extension User {
     @NSManaged public var created: String
     @NSManaged public var commentCount: Int64
     @NSManaged public var about: String?
-    @NSManaged public var followerCount: Int16
-    @NSManaged public var followingCount: Int16
+    @NSManaged public var followerCount: Int64
+    @NSManaged public var followingCount: Int64
     @NSManaged public var followingLimit: Int64
 
     @NSManaged public var owner: UserSecretOwnerKey?
